trian & test code split
코드 정상적으로 동작 확인 완료
